refactor(marge): split marge.go into spy/routes/responses/group files
Move-only commit 1 of catalog #26; Server struct + options stay in
marge.go, wire bytes pinned by the existing marge tests.
